+++ app/routers/user.py
@@ -4,6 +4,7 @@ from app.database import get_db
 from app import schemas, models
 from app.utils import hash
 from typing import List
+from app.oauth2 import get_current_user
 
 
 router = APIRouter(
@@ -31,21 +32,29 @@ async def show(id:int, db:Session=Depends(get_db)):
 
 
 @router.delete('/{id}')
-async def destroy(id:int, db:Session=Depends(get_db)):
+async def destroy(id:int, db:Session=Depends(get_db), current_user:int=Depends(get_current_user)):
     user_del = db.query(models.User).filter(models.User.id==id).first()
     if not user_del:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="No user available with id {id}")
+
+    if user_del.id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform actions")
+    
     db.delete(user_del)
     db.commit()
     return {'msg':'Deleted'}
 
 
 @router.put('/{id}')
-async def update(id:int, request:schemas.UserCreate, db:Session=Depends(get_db)):
+async def update(id:int, request:schemas.UserCreate, db:Session=Depends(get_db), current_user:int=Depends(get_current_user)):
     user_upd = db.query(models.User).filter(models.User.id==id)
     user = user_upd.first()
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="No user available with id {id}")
+
+    if user.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform actions")
+
     hashed_password = hash(request.password)
     request.password = hashed_password
     update_data = request.dict(exclude_unset=True)
